+++ packages/styled-components/components/link/index.js
@@ -5,7 +5,7 @@ import useClientNavigationOnClick from
 import useStandardProps from '@irvingjs/styled/hooks/useStandardProps';
 import {
   standardPropTypes,
-  standardDefaultProps,
+  getStandardDefaultProps,
 } from '@irvingjs/styled/types/propTypes';
 import * as defaultStyles from './themes/default';
 
@@ -58,7 +58,7 @@ const Link = (props) => {
 };
 
 Link.defaultProps = {
-  ...standardDefaultProps,
+  ...getStandardDefaultProps(),
   ariaHidden: null,
   theme: defaultStyles,
   onClick: false,