+++ packages/core/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0-alpha.14](https://github.com/alleyinteractive/irving/packages/core/compare/v4.2.0-alpha.13...v4.2.0-alpha.14) (2020-09-25)
+
+
+### Bug Fixes
+
+* **resolvecomponents:** remove console log ([003949b](https://github.com/alleyinteractive/irving/packages/core/commit/003949bbc2f3c7bcb6a5d8879a8aaa3deb556f04))
+* **sitetheme:** fix accidentally persisted/mutated default provider config ([bda394a](https://github.com/alleyinteractive/irving/packages/core/commit/bda394af249e2816e87cfd0c2012c2e63ef377e0))
+
+
+
+
+
 # [4.2.0-alpha.12](https://github.com/alleyinteractive/irving/packages/core/compare/v4.1.0...v4.2.0-alpha.12) (2020-09-24)